Support import-js@0.7.0
In version 0.7.0, ImportJS has moved to a fully-node-js framework,
instead of the core being written in Ruby. Additionally, the editor
plugins are broken out into separate repositories to make dependency
management easier.

This version of the import-js.el plugin makes use of a
globally-installed import-js npm package. Instead of running a Ruby
child process, we shell out to importjs on every call. This makes the
plugin quite a bit simpler, and we have the same level of functionality
as with the previous version, but there is now some noticeable lag when
performing import-js operations.

Normally, I don't do these *that* often, and the lag is quite small, so
maybe it isn't a problem.

The previous version of this plugin did not support multiple import
matches, and this version is no different. If multiple matches are
returned from the 'word' command, the import will silently fail.