Handle selecting multiple matches for 'word'
It is the case in many projects that an import could resolve to two or
more potential filepaths.

On the 'word' command, handle this case and prompt the user with a
list using grizzl, which is the base library used by fiplr, which I
currently use to great effect, a fuzzy-find package. The interface
is simple (using keyup/keydown) and will do the fuzzy matching for you.

If your project has no such multiple matches, there should be no net
effect, as there will not be a prompt and the import will complete.

The second step after enabling the 'word' command will be enabling the
multiple matching for 'fix', which will have to iterate over unresolved
imports.

I'm requiring the grizzl package the same way fiplr is, but I'm not
actually familiar with dependency management in MELPA, so I'll have to
test this out once it gets merged and MELPA picks up the change, which
is usually a few hours.