Use importjsd instead of importjs CLI
Version 2.1.0 of core import-js introduced some additional functionality
that only works in the daemon version. We also prefer the daemon process
version, because it is much faster (as it doesn't have to initialize a
node.js runtime). We can also take advantage of watchman if installed
locally.

This implementation is relatively straightforward, with one caveat: for
larger files, node.js sends chunks to STDOUT. We need to handle these
chunks in our process filter and only act when we see a newline.

This version of emacs-import-js removes support for using the import-js
CLI.