Save the buffer associated with the `importjs` command
When having a hook on window switch to save a file and having a file
to fix imports `after save`, the new file content was written to the
window I switched to due to the asynchronous call to `importjs`.

Now we save the buffer for which the command was invoked and write to
that buffer explicitly.